requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bidict==0.23.1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
dataclasses-json==0.6.7
datasets==3.5.0
dill==0.3.8
faiss-cpu==1.10.0
fastapi==0.115.12
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.2
granian==2.2.6
greenlet==3.2.2
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.31.1
idna==3.10
jinja2==3.1.6
joblib==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.24
langchain-community==0.3.22
langchain-core==0.3.56
langchain-huggingface==0.1.2
langchain-ollama==0.3.2
langchain-text-splitters==0.3.8
langsmith==0.3.37
mako==1.3.10
markdown-it-py==3.0.0
markupsafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
mypy-extensions==1.1.0
networkx==3.4.2
numpy==2.2.5
ollama==0.4.8
orjson==3.10.16
packaging==25.0
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.8
propcache==0.3.1
psutil==7.0.0
pyarrow==20.0.0
pydantic==2.11.4
pydantic-core==2.33.2
pydantic-settings==2.9.1
pygments==2.19.1
pyjwt==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.0
python-multipart==0.0.20
python-socketio==5.13.0
pytz==2025.2
pyyaml==6.0.2
redis==5.3.0
reflex==0.7.10.post1
reflex-hosting-cli==0.1.49
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
setuptools==79.0.1
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.40
sqlmodel==0.0.24
starlette==0.46.2
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.2.2
tqdm==4.67.1
transformers==4.51.3
typer==0.15.2
typing-extensions==4.13.2
typing-inspect==0.9.0
typing-inspection==0.4.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
watchfiles==1.0.5
wrapt==1.17.2
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0
